Chapter-14/requirements.txt
# requirements.txt
openai-agents
requests
python-slugify
rich
python-dotenv